Moved Doc type to selectDocsApi Moved /api/docs.ts to /preference/selectDocsApi.ts Added all source docs to redux
