+++ frontend/src/models/misc.ts
@@ -3,14 +3,3 @@ export type ActiveState = 'ACTIVE' | 'INACTIVE';
 export type User = {
   avatar: string;
 };
-
-export type Doc = {
-  name: string;
-  language: string;
-  version: string;
-  description: string;
-  fullName: string;
-  dat: string;
-  docLink: string;
-  model: string;
-};
